+++ crates/tui-realm-stdlib/examples/chart.rs
@@ -36,10 +36,7 @@ use rand::{rngs::ThreadRng, thread_rng, Rng};
 use std::thread::sleep;
 use std::time::{Duration, Instant};
 
-use tui_realm_stdlib::components::{
-    chart::{Chart, ChartPropsBuilder},
-    label,
-};
+use tui_realm_stdlib::{Chart, ChartPropsBuilder, Label, LabelPropsBuilder};
 use tuirealm::props::{
     borders::{BorderType, Borders},
     Alignment, Dataset,
@@ -161,8 +158,8 @@ fn main() {
     );
     myview.mount(
         COMPONENT_EVENT,
-        Box::new(label::Label::new(
-            label::LabelPropsBuilder::default()
+        Box::new(Label::new(
+            LabelPropsBuilder::default()
                 .with_foreground(Color::Cyan)
                 .with_text(String::from("Event will appear here"))
                 .build(),
@@ -235,11 +232,10 @@ impl Update for Model {
                 }
                 (component, event) => {
                     // Update span
-                    let props = label::LabelPropsBuilder::from(
-                        self.view.get_props(COMPONENT_EVENT).unwrap(),
-                    )
-                    .with_text(format!("{} => '{:?}'", component, event))
-                    .build();
+                    let props =
+                        LabelPropsBuilder::from(self.view.get_props(COMPONENT_EVENT).unwrap())
+                            .with_text(format!("{} => '{:?}'", component, event))
+                            .build();
                     // Report submit
                     let _ = self.view.update(COMPONENT_EVENT, props);
                     None