+++ crates/tui-realm-stdlib/examples/list.rs
@@ -33,7 +33,7 @@ use utils::keymap::*;
 use std::thread::sleep;
 use std::time::{Duration, Instant};
 
-use tui_realm_stdlib::components::{label, list};
+use tui_realm_stdlib::{Label, LabelPropsBuilder, List, ListPropsBuilder};
 use tuirealm::props::borders::{BorderType, Borders};
 use tuirealm::props::{Alignment, TableBuilder, TextSpan};
 use tuirealm::{Msg, PropsBuilder, Update, View};
@@ -90,8 +90,8 @@ fn main() {
     // Mount the component you need; we'll use a Label and an Input
     myview.mount(
         COMPONENT_TABLE,
-        Box::new(list::List::new(
-            list::ListPropsBuilder::default()
+        Box::new(List::new(
+            ListPropsBuilder::default()
                 .with_borders(Borders::ALL, BorderType::Thick, Color::Blue)
                 .with_title("My table data", Alignment::Center)
                 .with_rows(
@@ -150,8 +150,8 @@ fn main() {
     );
     myview.mount(
         COMPONENT_TABLE_2,
-        Box::new(list::List::new(
-            list::ListPropsBuilder::default()
+        Box::new(List::new(
+            ListPropsBuilder::default()
                 .with_borders(Borders::ALL, BorderType::Thick, Color::Blue)
                 .with_title("My table data", Alignment::Center)
                 .with_rows(
@@ -210,8 +210,8 @@ fn main() {
     );
     myview.mount(
         COMPONENT_SCROLLTABLE,
-        Box::new(list::List::new(
-            list::ListPropsBuilder::default()
+        Box::new(List::new(
+            ListPropsBuilder::default()
                 .with_borders(Borders::ALL, BorderType::Thick, Color::Blue)
                 .with_highlighted_str(Some("🚀"))
                 .with_highlighted_color(Color::LightBlue)
@@ -273,8 +273,8 @@ fn main() {
     );
     myview.mount(
         COMPONENT_SCROLLTABLE_2,
-        Box::new(list::List::new(
-            list::ListPropsBuilder::default()
+        Box::new(List::new(
+            ListPropsBuilder::default()
                 .with_borders(Borders::ALL, BorderType::Thick, Color::Blue)
                 .with_highlighted_str(Some("🚀"))
                 .with_max_scroll_step(4)
@@ -337,8 +337,8 @@ fn main() {
     );
     myview.mount(
         COMPONENT_EVENT,
-        Box::new(label::Label::new(
-            label::LabelPropsBuilder::default()
+        Box::new(Label::new(
+            LabelPropsBuilder::default()
                 .with_foreground(Color::Cyan)
                 .build(),
         )),
@@ -415,11 +415,10 @@ impl Update for Model {
                 }
                 (component, event) => {
                     // Update span
-                    let props = label::LabelPropsBuilder::from(
-                        self.view.get_props(COMPONENT_EVENT).unwrap(),
-                    )
-                    .with_text(format!("{} => '{:?}'", component, event))
-                    .build();
+                    let props =
+                        LabelPropsBuilder::from(self.view.get_props(COMPONENT_EVENT).unwrap())
+                            .with_text(format!("{} => '{:?}'", component, event))
+                            .build();
                     // Report submit
                     let _ = self.view.update(COMPONENT_EVENT, props);
                     None