+++ crates/tui-realm-stdlib/examples/paragraph.rs
@@ -33,7 +33,7 @@ use utils::keymap::*;
 use std::thread::sleep;
 use std::time::{Duration, Instant};
 
-use tui_realm_stdlib::components::{label, paragraph};
+use tui_realm_stdlib::{Label, LabelPropsBuilder, Paragraph, ParagraphPropsBuilder};
 use tuirealm::props::{Alignment, TextSpan};
 use tuirealm::{Msg, PropsBuilder, Update, View};
 // tui
@@ -87,8 +87,8 @@ fn main() {
     // Mount the component you need; we'll use a Label and an Input
     myview.mount(
         COMPONENT_PARAGRAPH,
-        Box::new(paragraph::Paragraph::new(
-            paragraph::ParagraphPropsBuilder::default()
+        Box::new(Paragraph::new(
+            ParagraphPropsBuilder::default()
                 .bold()
                 .with_background(Color::White)
                 .with_foreground(Color::Black)
@@ -103,8 +103,8 @@ fn main() {
     );
     myview.mount(
         COMPONENT_PARAGRAPH_2,
-        Box::new(paragraph::Paragraph::new(
-            paragraph::ParagraphPropsBuilder::default()
+        Box::new(Paragraph::new(
+            ParagraphPropsBuilder::default()
                 .bold()
                 .with_foreground(Color::LightBlue)
                 .with_text_alignment(Alignment::Center)
@@ -119,8 +119,8 @@ fn main() {
     );
     myview.mount(
         COMPONENT_EVENT,
-        Box::new(label::Label::new(
-            label::LabelPropsBuilder::default()
+        Box::new(Label::new(
+            LabelPropsBuilder::default()
                 .with_foreground(Color::Cyan)
                 .build(),
         )),
@@ -193,11 +193,10 @@ impl Update for Model {
                 }
                 (component, event) => {
                     // Update span
-                    let props = label::LabelPropsBuilder::from(
-                        self.view.get_props(COMPONENT_EVENT).unwrap(),
-                    )
-                    .with_text(format!("{} => '{:?}'", component, event))
-                    .build();
+                    let props =
+                        LabelPropsBuilder::from(self.view.get_props(COMPONENT_EVENT).unwrap())
+                            .with_text(format!("{} => '{:?}'", component, event))
+                            .build();
                     // Report submit
                     let _ = self.view.update(COMPONENT_EVENT, props);
                     None