+++ crates/tui-realm-stdlib/examples/textarea.rs
@@ -33,7 +33,7 @@ use utils::keymap::*;
 use std::thread::sleep;
 use std::time::{Duration, Instant};
 
-use tui_realm_stdlib::components::{label, textarea};
+use tui_realm_stdlib::{Label, LabelPropsBuilder, Textarea, TextareaPropsBuilder};
 use tuirealm::props::borders::{BorderType, Borders};
 use tuirealm::props::{Alignment, TextSpan};
 use tuirealm::{Msg, PropsBuilder, Update, View};
@@ -88,8 +88,8 @@ fn main() {
     // Mount the component you need; we'll use a Label and an Input
     myview.mount(
         COMPONENT_TEXTAREA,
-        Box::new(textarea::Textarea::new(
-            textarea::TextareaPropsBuilder::default()
+        Box::new(Textarea::new(
+            TextareaPropsBuilder::default()
                 .with_foreground(Color::White)
                 .italic()
                 .with_borders(Borders::ALL, BorderType::Rounded, Color::LightRed)
@@ -111,8 +111,8 @@ fn main() {
     );
     myview.mount(
         COMPONENT_TEXTAREA_2,
-        Box::new(textarea::Textarea::new(
-            textarea::TextareaPropsBuilder::default()
+        Box::new(Textarea::new(
+            TextareaPropsBuilder::default()
                 .with_foreground(Color::Black)
                 .with_foreground(Color::White)
                 .italic()
@@ -134,8 +134,8 @@ fn main() {
     );
     myview.mount(
         COMPONENT_EVENT,
-        Box::new(label::Label::new(
-            label::LabelPropsBuilder::default()
+        Box::new(Label::new(
+            LabelPropsBuilder::default()
                 .with_foreground(Color::Cyan)
                 .build(),
         )),
@@ -208,11 +208,10 @@ impl Update for Model {
                 }
                 (component, event) => {
                     // Update span
-                    let props = label::LabelPropsBuilder::from(
-                        self.view.get_props(COMPONENT_EVENT).unwrap(),
-                    )
-                    .with_text(format!("{} => '{:?}'", component, event))
-                    .build();
+                    let props =
+                        LabelPropsBuilder::from(self.view.get_props(COMPONENT_EVENT).unwrap())
+                            .with_text(format!("{} => '{:?}'", component, event))
+                            .build();
                     // Report submit
                     let _ = self.view.update(COMPONENT_EVENT, props);
                     None